[fix] Annotations on exception handler methods do not work
Fixes gh-150
